CONTRIB-254 Arcadia fixes (#3475)
* Macos: set ENABLE_EMBEDDED_COMPILER=0 by default (because contrib/llvm temporary broken under macos)

* \n

* arcadia fix

* fix arcadia link

* wip

* wip

* wip

* fix

* wip

* wip

* fix

* try use users from server dir

* Revert "try use users from server dir"

This reverts commit 79b2c6fd9c8d3d07cd3d06ed42abd49721563f88.

* Fix logging

* fix

* Update ZooKeeperImpl.h
