+++ dbms/src/Common/ZooKeeper/ZooKeeperImpl.h
@@ -240,4 +240,29 @@ private:
     CurrentMetrics::Increment active_session_metric_increment{CurrentMetrics::ZooKeeperSession};
 };
 
+struct ZooKeeperResponse;
+using ZooKeeperResponsePtr = std::shared_ptr<ZooKeeperResponse>;
+
+/// Exposed in header file for Yandex.Metrica code.
+struct ZooKeeperRequest : virtual Request
+{
+    ZooKeeper::XID xid = 0;
+    bool has_watch = false;
+    /// If the request was not send and the error happens, we definitely sure, that is has not been processed by the server.
+    /// If the request was sent and we didn't get the response and the error happens, then we cannot be sure was it processed or not.
+    bool probably_sent = false;
+
+    virtual ~ZooKeeperRequest() {}
+
+    virtual ZooKeeper::OpNum getOpNum() const = 0;
+
+    /// Writes length, xid, op_num, then the rest.
+    void write(WriteBuffer & out) const;
+
+    virtual void writeImpl(WriteBuffer &) const = 0;
+
+    virtual ZooKeeperResponsePtr makeResponse() const = 0;
+};
+
+
 }
